+++ token_supplies/CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.3](https://github.com/maidsafe/safe_network/compare/token_supplies-v0.1.2...token_supplies-v0.1.3) - 2024-01-24
+
+### Other
+- update dependencies
+
 ## [0.1.2](https://github.com/maidsafe/safe_network/compare/token_supplies-v0.1.1...token_supplies-v0.1.2) - 2024-01-23
 
 ### Other